pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai==1.12.1",
    "ipykernel>=7.2.0",
    "python-dotenv==1.1.0",
]
